# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kash-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "User settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSettingsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not logged in"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUserSettingsPayload": {
        "type": "object",
        "required": [
          "main_currency"
        ],
        "properties": {
          "main_currency": {
            "type": "string"
          }
        }
      },
      "UserSettings": {
        "type": "object",
        "required": [
          "main_currency"
        ],
        "properties": {
          "main_currency": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "axum_session"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}